[ffmemless] Use LIBS, not QMAKE_LFLAGS to link a library. Contributes to JB#34869.
The order of variables used is:
LFLAGS, SOURCES, LIBS.

In the case of -Wl,--as-needed, this could mean that the symbols used from the
library were dropped (because the sources using them hadn't yet been seen by the
linker).

This reared its head as a runtime load error on the ffmemless plugin.